Extract Iovec library from Uring and restructure repository
