+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "liburing"]
-	path = liburing
+	path = vendor/liburing
 	url = https://github.com/axboe/liburing
