--- dune
@@ -1,39 +1,0 @@
-(data_only_dirs liburing)
-
-(library
- (name            uring)
- (public_name     uring)
- (modules         region iovec uring)
- (libraries       bigstringaf)
- (foreign_archives uring)
- (foreign_stubs
-  (language c)
-  (flags (:standard -O0))
-  (names uring_stubs)
-  (include_dirs include)
-  (extra_deps include/liburing/compat.h)
- ))
-
-(rule
- (deps
-  (source_tree liburing))
- (targets liburing.a dlluring.so barrier.h compat.h io_uring.h liburing.h)
- (action
-  (no-infer
-   (progn
-    (chdir
-     liburing
-     (progn
-      (run ./configure)
-      ;; TODO discover.ml for liburing cflags
-      (setenv
-       CFLAGS
-       "-g -fomit-frame-pointer -O0 -fPIC"
-       (run make -j -C src))))
-    (copy liburing/src/liburing.a liburing.a)
-    (copy liburing/src/liburing.so.2.0.0 dlluring.so)
-    (copy liburing/src/include/liburing.h liburing.h)
-    (copy liburing/src/include/liburing/io_uring.h io_uring.h)
-    (copy liburing/src/include/liburing/barrier.h barrier.h)
-    (copy liburing/src/include/liburing/compat.h compat.h)
-))))
