+++ lib/iovec/dune
@@ -1,0 +1,6 @@
+(library
+ (name iovec)
+ (public_name iovec)
+ (foreign_stubs
+  (language c)
+  (names iovec_stubs)))
