+++ tests/basic_file_read.ml
@@ -3,9 +3,9 @@
 let () =
   let t = Uring.create ~queue_depth:1 ~default:() () in
   let fd = Unix.(handle_unix_error (openfile "test.txt" [O_RDONLY]) 0) in
-  let b1 = Uring.Iovec.alloc_buf 3 in
-  let b2 = Uring.Iovec.alloc_buf 7 in
-  let iov = Uring.Iovec.alloc [|b1;b2|] in
+  let b1 = Iovec.Buffer.create 3 in
+  let b2 = Iovec.Buffer.create 7 in
+  let iov = Iovec.alloc [|b1;b2|] in
   let r = Uring.readv t fd iov () in assert(r);
   let res = Uring.submit t in
   Printf.eprintf "submitted %d\n%!" res;
@@ -15,7 +15,7 @@ let () =
     | Some v -> v
   in
   let (), res = retry () in
-  Uring.Iovec.free iov;
+  Iovec.free iov;
   Printf.eprintf "res %d\n%!" res;
   Printf.eprintf "%s -- %s\n%!" (Bigstringaf.to_string b1) (Bigstringaf.to_string b2);
   let off = 3 in
