+++ tests/dune
@@ -1,34 +1,38 @@
 (library
-  (name urcp_lib)
-  (modules urcp_lib)
-  (libraries fmt unix uring logs))
+ (name urcp_lib)
+ (modules urcp_lib)
+ (libraries fmt unix uring logs))
 
 (library
-  (name urcp_fixed_lib)
-  (modules urcp_fixed_lib)
-  (libraries fmt unix uring logs))
+ (name urcp_fixed_lib)
+ (modules urcp_fixed_lib)
+ (libraries fmt unix uring logs))
 
 (library
-  (name lwtcp_lib)
-  (modules lwtcp_lib)
-  (libraries lwt.unix logs))
+ (name lwtcp_lib)
+ (modules lwtcp_lib)
+ (libraries lwt.unix logs))
 
 (executable
  (name urcat)
  (modules urcat)
  (public_name urcat)
+ (package uring)
  (libraries bigstringaf unix uring))
 
 (executable
  (name urcp)
  (modules urcp)
  (public_name urcp)
- (libraries cmdliner logs.cli logs.fmt fmt.tty fmt.cli urcp_fixed_lib urcp_lib))
+ (package uring)
+ (libraries cmdliner logs.cli logs.fmt fmt.tty fmt.cli urcp_fixed_lib
+   urcp_lib))
 
 (executable
  (name lwtcp)
  (modules lwtcp)
  (public_name lwtcp)
+ (package uring)
  (libraries cmdliner logs.cli logs.fmt fmt.tty fmt.cli lwtcp_lib))
 
 (executable
@@ -39,7 +43,8 @@
 (executable
  (name cptest)
  (modules cptest)
- (libraries urcp_fixed_lib urcp_lib lwtcp_lib uring bos bechamel notty.unix bechamel-notty))
+ (libraries urcp_fixed_lib urcp_lib lwtcp_lib uring bos bechamel notty.unix
+   bechamel-notty))
 
 (rule
  (alias runtest)
@@ -49,4 +54,5 @@
 (rule
  (alias runbenchmark)
  (deps urcp.exe)
- (action (run ./cptest.exe ./urcp.exe)))
+ (action
+  (run ./cptest.exe ./urcp.exe)))
